refactor order service
