+++ gateway/src/main/java/com/concordeu/gateway/GatewayApplicationConfig.java
@@ -1,6 +1,5 @@
-package com.concordeu.order.config;
+package com.concordeu.gateway;
 
-import com.google.gson.Gson;
 import io.github.resilience4j.circuitbreaker.CircuitBreakerConfig;
 import io.github.resilience4j.timelimiter.TimeLimiterConfig;
 import org.springframework.cloud.circuitbreaker.resilience4j.ReactiveResilience4JCircuitBreakerFactory;
@@ -11,16 +10,11 @@ import org.springframework.context.annotation.Configuration;
 import java.time.Duration;
 
 @Configuration
-public class OrderApplicationConfig {
+public class GatewayApplicationConfig {
     @Bean
-    public Gson gson() {
-        return new Gson();
-    }
-
-    @Bean
-    public Customizer<ReactiveResilience4JCircuitBreakerFactory> orderServiceCustomizer() {
+    public Customizer<ReactiveResilience4JCircuitBreakerFactory> orderServiceCusomtizer() {
         return factory -> factory.configure(builder -> builder
-                .timeLimiterConfig(TimeLimiterConfig.custom().timeoutDuration(Duration.ofSeconds(2)).build())
-                .circuitBreakerConfig(CircuitBreakerConfig.ofDefaults()), "order-service");
+                .timeLimiterConfig(TimeLimiterConfig.custom().timeoutDuration(Duration.ofSeconds(4)).build())
+                .circuitBreakerConfig(CircuitBreakerConfig.ofDefaults()), "ecommerceServices");
     }
 }
