+++ order/src/main/java/com/concordeu/order/service/ChargeServiceImpl.java
@@ -77,10 +77,10 @@ public class ChargeServiceImpl implements ChargeService {
                 .bodyValue(request)
                 .retrieve()
                 .bodyToMono(PaymentDto.class)
-                /*.transform(it ->
-                        reactiveCircuitBreakerFactory.create("charge-service")
+                .transform(it ->
+                        reactiveCircuitBreakerFactory.create("order-service")
                                 .run(it, throwable -> (Mono.just(PaymentDto.builder().username("Ooops...").build())))
-                )*/
+                )
                 .block();
     }
 
@@ -92,7 +92,7 @@ public class ChargeServiceImpl implements ChargeService {
                 .retrieve()
                 .bodyToMono(PaymentDto.class)
                 .transform(it ->
-                        reactiveCircuitBreakerFactory.create("charge-service")
+                        reactiveCircuitBreakerFactory.create("order-service")
                                 .run(it, throwable -> (Mono.just(PaymentDto.builder().username("Ooops...").build())))
                 )
                 .block();
