+++ order/src/main/resources/application-dev.yml
@@ -117,6 +117,18 @@ springdoc:
 kafka:
   bootstrapAddress: localhost:9092
 
+#Resilience4j
+resilience4j.circuitbreaker:
+  instances:
+    order-service:
+      failureRateThreshold: 50
+      minimumNumberOfCalls: 10
+      slidingWindowType: TIME_BASED
+      slidingWindowSize: 10
+      waitDurationInOpenState: 1000
+      permittedNumberOfCallsInHalfOpenState: 5
+      register-health-indicator: true
+
 catalog:
   base:
     uri: http://127.0.0.1:8084/api/v1
